+++ test/ttmlir/EmitC/TTNN/quantization/requantize.mlir
@@ -8,10 +8,10 @@
 #ttnn_layout1 = #ttnn.ttnn_layout<(d0, d1, d2, d3) -> (d0 * 960 + d1 * 320 + d2, d3), <1x1>, memref<30x10x!tt.tile<32x32, u8>, #dram>, <interleaved>>
 
 module {
-  func.func @forward(%arg0: tensor<1x3x320x320x!quant.uniform<i8:f32, 0.1>>) -> tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>> {
-    %0 = tensor.empty() : tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>
-    %1 = "ttir.requantize"(%arg0, %0) : (tensor<1x3x320x320x!quant.uniform<i8:f32, 0.1>>, tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>) -> tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>
+  func.func @forward(%arg0: tensor<1x3x320x320x!quant.uniform<i32:f32, 0.1>>) -> tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>> {
+    %0 = tensor.empty() : tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>
+    %1 = "ttir.requantize"(%arg0, %0) : (tensor<1x3x320x320x!quant.uniform<i32:f32, 0.1>>, tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>) -> tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>
     // CHECK: "ttnn.requantize"
-    return %1 : tensor<1x3x320x320x!quant.uniform<i8:f32, 0.2>>
+    return %1 : tensor<1x3x320x320x!quant.uniform<i32:f32, 0.2>>
   }
 }
